testing: add glob-test crate and port tests
Before starting the rewrite, it's useful to migrate to a more flexible
testing infrastructure first. We do this first, and ensure the existing
implementation passes the test suite. The new test infrastructure comes
in the form of the glob-test crate, which provides a way to read tests
from TOML files and execute them.

This process did uncover one bug in the existing implementation. Namely,
that `[.-0]` matches `/` even when `literal_separator` is enabled.