readme: add disclaimer
I noticed that there have been a couple mistaken pull requests here
and unfortunately Google still lists this as the third hit (and first
GitHub hit) for the search "rust globset". It'd be nice to include a
simple disclaimer at the top of the README directing people to the real
maintained crate.